Rename server runtime package to bootstrap
